extracted createSuspendedSale
